Lo que podemos recuperar sin queries, y correcion de tipo de dato en el AbstracDao las implementaciones estaban con int y el AbstracDao con long

+++ otra-reunion-mas/otra-reunion-mas/src/main/java/com/linkedin/learning/otrareunionmas/dao/Dao.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  * 
  */
 public interface Dao<T> {
-	Optional<T> get(long id);
+	Optional<T> get(int id);
 
 	List<T> getAll();
